+++ packages/cortex-core/src/components/input/input.ts
@@ -2,7 +2,7 @@ import { ClassProp, tv, VariantProps } from 'tailwind-variants';
 import { clsx } from 'clsx';
 
 export const inputBaseStyles = tv({
-  base: 'relative min-h-[44px] h-auto flex px-centi py-1.5 focus-within:outline focus-within:outline-1 text-base font-bold items-center bg-transparent border rounded-mili',
+  base: 'relative min-h-[44px] h-auto flex px-centi py-1.5 bg-white focus-within:outline focus-within:outline-1 text-base font-bold items-center border rounded-mili has-[:disabled]:bg-transparent',
   variants: {
     intent: {
       default: [
@@ -37,13 +37,13 @@ export const inputBaseStyles = tv({
 });
 
 const labelBaseStyles = tv({
-  base: 'absolute duration-300 -translate-y-2 scale-[0.72] top-2.5 z-10 origin-[0] peer-placeholder-shown:scale-100 peer-placeholder-shown:-translate-y-0 peer-focus:scale-[0.72] peer-focus:-translate-y-2 peer-focus:top-2.5 font-bold text-base peer-disabled:text-secondary-light select-none',
+  base: 'absolute duration-300 -translate-y-2 scale-[0.72] top-2.5 z-10 origin-[0] peer-placeholder-shown:scale-100 peer-placeholder-shown:-translate-y-0 peer-focus:scale-[0.72] peer-focus:-translate-y-2 peer-focus:top-2.5 font-bold text-base peer-disabled:text-secondary-light select-none w-max',
   variants: {
     intent: {
       default: [
         'text-secondary-medium',
         'peer-focus:text-secondary-medium',
-        'peer-placeholder-shown:text-secondary-xdark',
+        'peer-placeholder-shown:text-secondary-medium',
       ],
       success: ['text-success-medium'],
       warning: ['text-warning-medium'],
@@ -75,7 +75,7 @@ export const inputBox = (
 ) =>
   clsx(
     className,
-    'peer relative outline-none placeholder:text-base placeholder:text-secondary-xdark placeholder-shown:top-0 disabled:bg-opacity-0 disabled:placeholder:text-secondary-light bg-transparent',
+    'peer relative outline-none placeholder:text-base placeholder:text-secondary-medium placeholder-shown:top-0 disabled:placeholder:text-secondary-light disabled:text-secondary-light bg-transparent',
     {
       'focus:top-1.5': !placeholder && !!label,
       'top-1.5': !!label,